\name{asremlPlus-deprecated}
\alias{asremlPlus-deprecated}
\alias{addrm.terms.asreml}
\alias{choose.model.asreml}
\alias{recalc.wald.tab.asreml}
\alias{rmboundary.asreml}
\alias{sig.devn.reparam.asreml}
\alias{testranfix.asreml}
\alias{testrcov.asreml}
\alias{testswapran.asreml}
\alias{info.crit}
\alias{reml.lrt}
\title{Deprecated Functions in Package asremlPlus}
\description{These functions have been renamed and deprecated in \code{asremlPlus}: addrm.terms.asreml replaced by \code{\link{addrm.terms.asrtests}}, choose.model.asreml by \code{\link{addrm.terms.asrtests}}, recalc.wald.tab.asreml by \code{\link{recalc.wald.tab.asrtests}}, rmboundary.asreml by \code{\link{rmboundary.asrtests}}, sig.devn.reparam.asreml by \code{\link{sig.devn.reparam.asrtests}}, testranfix.asreml by \code{\link{testranfix.asrtests}}, testrcov.asreml by \code{\link{testrcov.asrtests}}, testswapran.asreml by \code{\link{testswapran.asrtests}}, info.crit by \code{\link{info.crit.asreml}}, reml.lrt by \code{\link{reml.lrt.asreml}}}
\usage{addrm.terms.asreml(...)
choose.model.asreml(...)
recalc.wald.tab.asreml(...)
rmboundary.asreml(...)
sig.devn.reparam.asreml(...)
testranfix.asreml(...)
testrcov.asreml(...)
testswapran.asreml(...)
info.crit(...)
reml.lrt(...)
}
\arguments{
  \item{...}{absorbs arguments passed from the old functions of the style foo.bar().}
}
\author{Chris Brien}
