\name{Oats.dat}
\docType{data}
\alias{Oats.dat}
\title{Data for an experiment to investigate nitrogen response of 3 oats varieties}
\description{
Yates (1937) describes a split-plot experiment that investigates the effects of three varieties of oats and four levels of Nitrogen fertilizer. The varieties are assigned to the main plots using a randomized complete block design with 6 blocks and the nitrogen levels are randomly assigned  to the subplots in each main plot.

The columns in the data frame are: Blocks, Wplots, Subplots, Variety, Nitrogen, xNitrogen, Yield. The column xNitrogen is a numeric version of the factor Nitrogen.
The response variable is Yield.

}
\usage{data(Oats.dat)}
\format{A data.frame containing 72 observations of 7 variables.}
\source{
Yates, F. (1937). The Design and Analysis of Factorial Experiments. \emph{Imperial Bureau of Soil Science, Technical Communication}, \bold{35}, 1-95. 
}
\author{Chris Brien}
\keyword{datasets}