\name{asremlPlusTips}
\alias{asremlPlusTips}
\title{The randomly-presented, startup tips.}
\description{The intermittent, randomly-presented, startup tips.}

\section{Startup tips}{
Need help? The manual is in the doc subdirectory of the package's install directory.

Find out what has changed in asremlPlus: enter news(package = 'asremlPlus').

Need help getting started? Look at the example in ?`asremlPlus-package`.

To avoid start-up message that ASReml-R is needed, load asreml before asremlPlus.

The methods for alldiffs and data.frame do not require asreml

Use suppressPackageStartupMessages() to eliminate all package startup messages.

To see all the intermittent, randomly-presented, startup tips enter ?asremlPlusTips.

To install the latest version: go to \url{http://chris.brien.name/rpackages}.

For versions between CRAN releases (and more) go to \url{http://chris.brien.name/rpackages}.
}
\author{Chris Brien}
\keyword{asreml}