\name{WaterRunoff.dat}
\docType{data}
\alias{WaterRunoff.dat}
\title{Data for an experiment to investigate the quality of water runoff over time}
\description{
This data is from an experiment to investigate the quality of water runoff. However, it has been modified to hide the true identity of the Species and Sources. It is used to provide executable examples of the functions listed under \bold{Examples}.
}
\usage{data(WaterRunoff.dat)}
\format{A data.frame containing 440 observations of 13 variables.}
\source{Kazemi, F. (pers. comm.)}
\author{Chris Brien}
\seealso{\code{\link{chooseModel.asrtests}}, \code{\link{reparamSigDevn.asrtests}}, \cr
         \code{\link{plotPredictions.data.frame}}, \code{\link{predictPlus.asreml}}, 
         \code{\link{predictPresent.asreml}}}
\keyword{datasets}